+++ workload_gen/inc/zipfian_generator.h
@@ -61,7 +61,7 @@ private:
 
     double Eta()
     {
-        return (1 - std::pow(2.0 / num_items_, 1 - theta_)) / (1 - zeta_2_ / zeta_n_);
+        return (1 - std::pow(2.0 / static_cast<double>(num_items_), 1 - theta_)) / (1 - zeta_2_ / zeta_n_);
     }
 
     ///
@@ -116,7 +116,7 @@ inline uint64_t ZipfianGenerator::Next(uint64_t num)
         return last_value_ = 1;
     }
 
-    return last_value_ = base_ + num * std::pow(eta_ * u - eta_ + 1, alpha_);
+    return last_value_ = static_cast<uint64_t>(static_cast<double>(base_) + static_cast<double>(num) * std::pow(eta_ * u - eta_ + 1, alpha_));
 }
 
 inline uint64_t ZipfianGenerator::Last()
